Re-org shared code around packages; Use a virtual environment
This moves the code that was shared into packages, which can be
installed editable, which is nice for DRY and automatically reflecting
code changes.

Because those shared code packages should not be installed system-wide,
and because virtual environments are fun, use a virtual environment too.

The bootstrap.py script helps people get set up. There's still a very
small amount of !DRY in that script, but meh :-).

Giant-building works; AudioQuake-building is borked due to paths, but
wanted to commit this as a continue point.

The clean.sh script was amended so as not to remove the virtual
environment.