Support Open Quartz in the Launcher
This will use Open Quartz if the registered Quake data can't be found,
except if the player has specifically asked to play Quake, in which case
an error message will be shown if the data can't be found.

To address: how to let the player know that Open Quartz will be used and
they're not actually playing Quake? Something for first-time start-up
(as well as the manual)?