[LDL] Fix maps with incorrect .wad paths during build
This takes the .map file as input and (if needed) replaces the path to
the .wad file with the correct one on the current system. The .map file
is saved as a .temp file (which is later removed) if changed. QBSP can
take the .temp file as input, so it fits nicely.