docs: Info for Mac users when App Translocation happens (#56)
The reason why the downloaded app was not running is because App Translocation was placing it in a read-only filesystem at the start. Found out from https://github.com/portacle/portacle/issues/22 (thanks!) that one way to address this is for the user to move the app (after they have decided it is for them, and granted it permission to open via Gatekeeper). Included some documentation in the .app (that is only apparent _after_ the Gatekeeper part, of course) to explain they need to move the app to make it work.

I had started another branch in which the data were kept outside of the app (and other path-related improvements), but sticking with this for now.